chore: reorganize function files into separate files for each function
This commit moves the functions `dateFormat`, `formatMinutes`, `getAge`, `numberFormat`, `objectKeys`, and `firstLetterToUppercase` into their own separate files in the `src/utils/function` directory. This makes it easier to find and modify specific functions, and also makes the codebase more modular. The `doesFolderExist` function has also been moved into its own file in the `src/utils/function/folder.ts` file.
