+++ src/commands/anniversaire/[sub-commands]/list.cmd.ts
@@ -3,7 +3,7 @@ import { getBirthdays } from "$core/api/requests/member";
 import { memberPerPage } from "./list.const";
 import { DayJS } from "$core/configs/day-js";
 import { simpleEmbed } from "$core/utils/discord/embed";
-import { dateFormat, getAge } from "$core/utils/function";
+import { dateFormat, getAge } from "$core/utils/function/date";
 import { msgParams } from "$core/utils/message";
 import { gqlRequest } from "$core/utils/request";
 import { commands } from "$core/configs/message/command";
