+++ src/utils/function/date.ts
@@ -1,11 +1,5 @@
 import type { Dayjs } from "dayjs";
-import { global } from "$core/configs/global";
 import { DayJS } from "$core/configs/day-js";
-import { existsSync, statSync } from "fs";
-
-export const numberFormat = (number: number): string => {
-  return number.toLocaleString(global.localFormat);
-};
 
 export const dateFormat = (date: Dayjs, separator = "-"): string => {
   return date.format(`DD[${separator}]MM[${separator}]YYYY`);
@@ -29,16 +23,4 @@ export const formatMinutes = (minutes: number): string => {
 
 export const getAge = (birth: Dayjs): number => {
   return DayJS().tz().diff(birth, "year");
-};
-
-export const firstLetterToUppercase = (string: string): string => {
-  return string.charAt(0).toUpperCase() + string.slice(1);
-};
-
-export const doesFolderExist = (path: string): boolean => {
-  return existsSync(path) && statSync(path).isDirectory();
-};
-
-export const objectKeys = <T extends object>(obj: T): (keyof T)[] => {
-  return Object.keys(obj) as (keyof T)[];
 };
