+++ src/utils/function/folder.ts
@@ -1,0 +1,5 @@
+import { existsSync, statSync } from "fs";
+
+export const doesFolderExist = (path: string): boolean => {
+  return existsSync(path) && statSync(path).isDirectory();
+};
