+++ src/utils/function/number.ts
@@ -1,0 +1,5 @@
+import { global } from "$core/configs/global";
+
+export const numberFormat = (number: number): string => {
+  return number.toLocaleString(global.localFormat);
+};
