+++ src/utils/function/object.ts
@@ -1,0 +1,3 @@
+export const objectKeys = <T extends object>(obj: T): (keyof T)[] => {
+  return Object.keys(obj) as (keyof T)[];
+};
