+++ src/utils/function/string.ts
@@ -1,0 +1,3 @@
+export const firstLetterToUppercase = (string: string): string => {
+  return string.charAt(0).toUpperCase() + string.slice(1);
+};
